styled the reply containers and post containers to make sure they fit the page, deleted unused variables, made styling consistent accross components
