+++ src/App.scss
@@ -2,37 +2,39 @@
   text-align: center;
   padding-top: 11%;
   padding-bottom: 12%;
+  // height: 100vh;
+  // min-height: 100vh;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
+// .App-logo {
+//   animation: App-logo-spin infinite 20s linear;
+//   height: 40vmin;
+//   pointer-events: none;
+// }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
+// .App-header {
+//   background-color: #282c34;
+//   min-height: 100vh;
+//   display: flex;
+//   flex-direction: column;
+//   align-items: center;
+//   justify-content: center;
+//   font-size: calc(10px + 2vmin);
+//   color: white;
+// }
 
-.App-link {
-  color: #61dafb;
-}
+// .App-link {
+//   color: #61dafb;
+// }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
+// @keyframes App-logo-spin {
+//   from {
+//     transform: rotate(0deg);
+//   }
+//   to {
+//     transform: rotate(360deg);
+//   }
+// }
 
 .h1 {
   color: #1b4570;
@@ -40,6 +42,7 @@
 .search-form {
   display: flex;
   flex-direction: column;
+  align-items: center;
   margin: 0 auto;
   width: 90%;
   position: relative;
@@ -53,12 +56,12 @@
   }
   .results-container {
     display: flex;
-    width: 100%;
+    width: 95.5%;
     flex-direction: column;
     position: absolute;
     z-index: 1;
     background-color: white;
-    margin-top: 12.5%;
+    margin-top: 18.5%;
     .suggestion-active {
       background-color: lightgoldenrodyellow;
     }
