+++ src/components/groups/GroupContainer.js
@@ -9,47 +9,53 @@ import MyAllegianceGroups from "../profile/MyAllegianceGroups";
 import NearbyGroups from "./NearbyGroups";
 
 function GroupContainer() {
-	const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
+  const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
 
-	return (
-		<div>
-			<SearchBar />
-			<Divider />
-			<MyGroups>
-				<GroupTitleHolder>
-					<H3>MY GROUPS</H3>
-				</GroupTitleHolder>
-				<>
-					<MyAllegianceGroups content={loggedInGroups} type={"groups"} />
-				</>
-			</MyGroups>
-			<Divider />
-			<H3>GROUPS NEAR YOU</H3>
-			<NearbyGroups />
-			<Divider />
-			<H3>DISCOVER</H3>
-			<GroupList />
-		</div>
-	);
+  return (
+    <Container>
+      <SearchBar />
+      <Divider />
+      <MyGroups>
+        <GroupTitleHolder>
+          <H3>MY GROUPS</H3>
+        </GroupTitleHolder>
+        <>
+          <MyAllegianceGroups content={loggedInGroups} type={"groups"} />
+        </>
+      </MyGroups>
+      <Divider />
+      <H3>GROUPS NEAR YOU</H3>
+      <NearbyGroups />
+      <Divider />
+      <H3>DISCOVER</H3>
+      <GroupList />
+    </Container>
+  );
 }
 export default GroupContainer;
 
+const Container = styled.div`
+  h3 {
+    text-align: left;
+    margin-left: 5%;
+  }
+`;
 const MyGroups = styled.div`
-	margin: 2% 5%;
+  margin: 2% 5%;
 `;
 
 const GroupTitleHolder = styled.div`
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-start;
-	align-items: center;
-	h3 {
-		margin: 0 1%;
-	}
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  h3 {
+    margin: 0 1%;
+  }
 `;
 
 const H3 = styled.h3`
-	font-size: 1rem;
-	font-weight: bold;
-	margin: 1% 0;
+  font-size: 1rem;
+  font-weight: bold;
+  margin: 1% 0;
 `;
