+++ src/components/groups/GroupList.js
@@ -10,71 +10,70 @@ import styled from "styled-components";
 import GroupCard from "./GroupCard";
 
 const GroupList = () => {
-	const [data, setData] = useState({ groups: [] });
+  const [data, setData] = useState({ groups: [] });
 
-	// Fetches Auth0 token for axios call
-	const [token] = useGetToken();
+  // Fetches Auth0 token for axios call
+  const [token] = useGetToken();
 
-	// Fetches user groups from Redux
-	const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
+  // Fetches user groups from Redux
+  const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
 
-	useEffect(() => {
-		const fetchData = async () => {
-			if (token) {
-				const groups = await axiosWithAuth([token]).post(`/groups/search`, {
-					column: "group_name",
-					row: ""
-				});
-				console.log("GroupsList data", groups.data);
+  useEffect(() => {
+    const fetchData = async () => {
+      if (token) {
+        const groups = await axiosWithAuth([token]).post(`/groups/search`, {
+          column: "group_name",
+          row: ""
+        });
 
-				// Get array of ids for groups the user already is a member of
-				const loggedInIDs = loggedInGroups.map(group => group.id);
+        // Get array of ids for groups the user already is a member of
+        const loggedInIDs = loggedInGroups.map(group => group.id);
 
-				const uniqueGroups = groups.data.groupByFilter.filter(
-					group => !loggedInIDs.includes(group.id)
-				);
+        const uniqueGroups = groups.data.groupByFilter.filter(
+          group => !loggedInIDs.includes(group.id)
+        );
 
-				uniqueGroups.sort((a, b) => b.members.length - a.members.length);
+        uniqueGroups.sort((a, b) => b.members.length - a.members.length);
 
-				setData({ groups: uniqueGroups });
-			}
-		};
+        setData({ groups: uniqueGroups });
+      }
+    };
 
-		fetchData();
-	}, [token, loggedInGroups]);
+    fetchData();
+  }, [token, loggedInGroups]);
 
-	if (!data) {
-		return (
-			<Loader active size="large">
-				Loading
-			</Loader>
-		);
-	}
-	//Component should only show top 20 , load more button below. Should be sortable by recent activity/group size/allegiances
+  if (!data) {
+    return (
+      <Loader active size="large">
+        Loading
+      </Loader>
+    );
+  }
+  //Component should only show top 20 , load more button below. Should be sortable by recent activity/group size/allegiances
 
-	return (
-		<SectionContainer>
-			<GroupListContainer>
-				{data.groups.map(group => {
-					return <GroupCard group={group} key={group.id} />;
-				})}
-			</GroupListContainer>
-		</SectionContainer>
-	);
+  return (
+    <SectionContainer>
+      <GroupListContainer>
+        {data.groups.map(group => {
+          return <GroupCard group={group} key={group.id} />;
+        })}
+      </GroupListContainer>
+    </SectionContainer>
+  );
 };
 
 const SectionContainer = styled.div`
-	display: flex;
-	flex-direction: column;
+  display: flex;
+  flex-direction: column;
 `;
 
 const GroupListContainer = styled.div`
-	display: flex;
-	flex-direction: row;
-	flex-wrap: wrap;
-	justify-content: center;
-	width: 100%;
-	margin-top: 1vh;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  width: 100%;
+  margin-top: 1vh;
 `;
 
 export default GroupList;
