+++ src/components/posts/PostsContainer.js
@@ -1,64 +1,43 @@
 import React, { useState, useEffect } from "react";
-import { useSelector } from "react-redux";
 import { axiosWithAuth } from "../utils/axiosWithAuth";
 import useGetToken from "../utils/useGetToken";
 
-import PostCard from "./PostCard";
 import PostForm from "./PostForm";
 
 import styled from "styled-components";
-import MatUiPostCard from "./MatUiPostCard";
+import PostCard from "./PostCard";
 
 const PostsContainer = props => {
   // Fetches Auth0 token for axios call
   const [token] = useGetToken();
-
-  const loggedInPosts = useSelector(state => state.userReducer.loggedInPosts);
   const [posts, setPosts] = useState([]);
   const [submitted, setSubmitted] = useState(false);
 
   useEffect(() => {
     const fetchData = async () => {
       const id = props.groupId;
-      if (token) {
-        const posts = await axiosWithAuth([token]).get(`/posts/group/${id}`);
-        console.log("%c Posts List", "color: orange; font-weight: bold;");
-        console.log(posts.data);
-        setPosts(posts.data.postsLoaded);
-        setSubmitted(false);
-      }
+      if (token)
+        try {
+          const posts = await axiosWithAuth([token]).get(`/posts/group/${id}`);
+          setPosts(posts.data.postsLoaded);
+          setSubmitted(false);
+        } catch {
+          setPosts([]);
+          setSubmitted(false);
+        }
     };
-
     fetchData();
-  }, [token, submitted]);
+  }, [token, submitted, props.groupId]);
 
-  const deletePost = async postId => {
-    const post = await axiosWithAuth([token]).delete(`/posts/${postId}`);
-    if (post) {
-      setSubmitted(true);
-    }
-  };
   return (
     <PostsWrapper>
       <PostListContainer>
         {posts.map(post => {
           return (
-            // <PostCard
-            //   post={post}
-            //   key={post.id}
-            //   deletePost={deletePost}
-            //   setSubmitted={setSubmitted}
-            // />
-            <MatUiPostCard
-              post={post}
-              key={post.id}
-              deletePost={deletePost}
-              setSubmitted={setSubmitted}
-            />
+            <PostCard post={post} key={post.id} setSubmitted={setSubmitted} />
           );
         })}
       </PostListContainer>
-
       <PostForm setSubmitted={setSubmitted} groupId={props.groupId} />
     </PostsWrapper>
   );
@@ -70,6 +49,7 @@ const PostListContainer = styled.div`
   flex-direction: column;
   align-items: center;
   background-color: #dee4e7;
+  padding-bottom: 15%;
 `;
 const PostsWrapper = styled.div`
   background-color: #dee4e7;
