+++ src/components/profile/Profile.js
@@ -7,155 +7,164 @@ import MyAllegianceGroups from "./MyAllegianceGroups";
 import { axiosWithAuth } from "../utils/axiosWithAuth";
 import useGetToken from "../utils/useGetToken";
 import { GET_GROUPS } from "../../reducers/userReducer";
-import defaultBanner from '../../assets/defaultBanner.jpg'
+import defaultBanner from "../../assets/defaultBanner.jpg";
 
 const Profile = props => {
-	const loggedInUser = useSelector(state => state.userReducer.loggedInUser);
-	const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
-	/*const loggedInAllegiances = useSelector(
+  const loggedInUser = useSelector(state => state.userReducer.loggedInUser);
+  const loggedInGroups = useSelector(state => state.userReducer.loggedInGroups);
+  /*const loggedInAllegiances = useSelector(
 		state => state.userReducer.loggedInAllegiances
 	);*/
-	const loggedInPosts = useSelector(state => state.userReducer.loggedInPosts);
-	const dispatch = useDispatch();
+  const loggedInPosts = useSelector(state => state.userReducer.loggedInPosts);
+  const dispatch = useDispatch();
 
-	//Fetches Auth0 token for axios call
-	const [token] = useGetToken();
+  //Fetches Auth0 token for axios call
+  const [token] = useGetToken();
 
-	useEffect(() => {
-		if (loggedInUser) {
-			const fetchData = async () => {
-				if (token) {
-					try {
-						const groups = await axiosWithAuth([token]).get(
-							`/groups_users/search/${loggedInUser.id}`
-						);
-						const userGroups = groups.data.groups.map(group => {
-							return {
-								name: group.group_name,
-								image: group.group_image,
-								id: group.group_id,
-								user_type: group.user_type
-							}
-						})
-						dispatch({ type: GET_GROUPS, payload: userGroups })
-					}
-					catch {
-						dispatch({ type: GET_GROUPS, payload: [] })
-					}
-				}
-			};
+  useEffect(() => {
+    if (loggedInUser) {
+      const fetchData = async () => {
+        if (token) {
+          try {
+            const groups = await axiosWithAuth([token]).get(
+              `/groups_users/search/${loggedInUser.id}`
+            );
+            const userGroups = groups.data.groups.map(group => {
+              return {
+                name: group.group_name,
+                image: group.group_image,
+                id: group.group_id,
+                user_type: group.user_type
+              };
+            });
+            dispatch({ type: GET_GROUPS, payload: userGroups });
+          } catch {
+            dispatch({ type: GET_GROUPS, payload: [] });
+          }
+        }
+      };
 
-			fetchData();
-		}
-	}, [token, loggedInUser, dispatch]);
+      fetchData();
+    }
+  }, [token, loggedInUser, dispatch]);
 
-	if (!loggedInUser) {
-		return (
-			<Loader active size='large'>Loading</Loader>
-		)
-	}
+  if (!loggedInUser) {
+    return (
+      <Loader active size="large">
+        Loading
+      </Loader>
+    );
+  }
 
-	return (
-		<ProfileContainer>
-			<div style={{ maxWidth: "100%" }}>
-				<Banner>
-					<BannerImage src={loggedInUser.banner_image || defaultBanner} fluid />
-				</Banner>
-				<ImageCrop>
-					{loggedInUser.image
-						? <ProfileImage src={loggedInUser.image} alt="Profile" />
-						: <Icon name="football ball" size="huge" circular style={{ fontSize: "5.3rem" }}
-						/>}
-				</ImageCrop>
-				<InfoHolder>
-					<Name>
-						{loggedInUser.first_name
-							? <h1>{`${loggedInUser.first_name} ${loggedInUser.last_name}`}</h1>
-							: null}
-						{props.match.url === "/profile" ? (
-							<Link to="/makeprofile">
-								<Icon name="edit outline" />
-							</Link>
-						) : null}
-					</Name>
-					<p>{loggedInUser.bio}</p>
-					<>
-						<H3>MY GROUPS</H3>
-						<MyAllegianceGroups content={loggedInGroups || []} type={"groups"} />
-					</>
-				</InfoHolder>
-				<div>
-					<PostHeader>
-						<H3>POSTS</H3>
-						<H3 style={{ color: "lightgrey" }} onClick={() => console.log(10)}>
-							{" "}
-							MANAGE POSTS
-						</H3>
-					</PostHeader>
-					<div>
-						{loggedInPosts ? (
-							loggedInPosts
-						) : (
-								<NoPosts>You haven't posted yet!</NoPosts>
-							)}
-					</div>
-				</div>
-			</div>
-		</ProfileContainer>
-	);
+  return (
+    <ProfileContainer>
+      <div style={{ maxWidth: "100%" }}>
+        <Banner>
+          <BannerImage src={loggedInUser.banner_image || defaultBanner} fluid />
+        </Banner>
+        <ImageCrop>
+          {loggedInUser.image ? (
+            <ProfileImage src={loggedInUser.image} alt="Profile" />
+          ) : (
+            <Icon
+              name="football ball"
+              size="huge"
+              circular
+              style={{ fontSize: "5.3rem" }}
+            />
+          )}
+        </ImageCrop>
+        <InfoHolder>
+          <Name>
+            {loggedInUser.first_name ? (
+              <h1>{`${loggedInUser.first_name} ${loggedInUser.last_name}`}</h1>
+            ) : null}
+            {props.match.url === "/profile" ? (
+              <Link to="/makeprofile">
+                <Icon name="edit outline" />
+              </Link>
+            ) : null}
+          </Name>
+          <p>{loggedInUser.bio}</p>
+          <>
+            <H3>MY GROUPS</H3>
+            <MyAllegianceGroups
+              content={loggedInGroups || []}
+              type={"groups"}
+            />
+          </>
+        </InfoHolder>
+        <div>
+          <PostHeader>
+            <H3>POSTS</H3>
+            <H3 style={{ color: "lightgrey" }} onClick={() => console.log(10)}>
+              {" "}
+              MANAGE POSTS
+            </H3>
+          </PostHeader>
+          <div>
+            {loggedInPosts ? (
+              loggedInPosts
+            ) : (
+              <NoPosts>You haven't posted yet!</NoPosts>
+            )}
+          </div>
+        </div>
+      </div>
+    </ProfileContainer>
+  );
 };
 
 const ProfileContainer = styled.div`
-	display: flex;
-	justify-content: center;
-	margin-top: -4.5%;
-	margin-bottom: 15%;
+  display: flex;
+  justify-content: center;
+  margin-top: -4.5%;
+  margin-bottom: 15%;
 `;
 
 const Banner = styled.div`
-	border-bottom: 10px solid black;
+  border-bottom: 10px solid black;
 `;
 
-const BannerImage = styled(Image)`
-`;
+const BannerImage = styled(Image)``;
 
 const InfoHolder = styled.div`
-	margin-top: 5%;
-	padding-bottom: 1rem;
-	border-bottom: 0.5px solid lightgray;
+  margin-top: 5%;
+  padding-bottom: 1rem;
+  border-bottom: 0.5px solid lightgray;
 `;
 
 const Name = styled.div`
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
 `;
 
 const ImageCrop = styled.div`
-	width: 150px;
-	height: 150px;
-	position: relative;
-	overflow: hidden;
-	border-radius: 50%;
-	border: 1px solid white;
-	margin-top: -8rem;
-	margin-left: auto;
-	margin-right: auto;
-	background-color: white;
+  width: 150px;
+  height: 150px;
+  position: relative;
+  overflow: hidden;
+  border-radius: 50%;
+  border: 1px solid white;
+  margin-top: -8rem;
+  margin-left: auto;
+  margin-right: auto;
+  background-color: white;
 `;
 const ProfileImage = styled.img`
-	display: inline;
-	margin-left: -25%; //centers the image
-	height: 100%;
-	width: auto;
-	max-width: none;
+  display: inline;
+  margin-left: -25%; //centers the image
+  height: 100%;
+  width: auto;
+  max-width: none;
 `;
 
 const H3 = styled.h3`
-	font-size: 1rem;
-	font-weight: bold;
-	margin-top: 0;
-	margin-bottom: 0;
+  font-size: 1rem;
+  font-weight: bold;
+  margin-top: 0;
+  margin-bottom: 0;
 `;
 
 const PostHeader = styled.div`
@@ -168,9 +177,9 @@ const PostHeader = styled.div`
 	margin-top: .5rem;`;
 
 const NoPosts = styled.p`
-	color: lightgrey;
-	font-weight: bold;
-	font-size: 1rem;
+  color: lightgrey;
+  font-weight: bold;
+  font-size: 1rem;
 `;
 
 export default Profile;
