+++ src/styled/Replies.scss
@@ -48,8 +48,8 @@
   display: flex;
   flex-direction: row;
   justify-content: center;
-  position: sticky;
-  bottom: 45px;
+  position: fixed;
+  bottom: 6.5%;
   width: 100%;
   background-color: #dee4e7;
   align-items: center;
